+++ watcher/github-release/pkg/taskpublisher.go
@@ -87,7 +87,11 @@ func buildFrontmatter(release Release, taskIDStr string, cfg TaskConfig) agentli
 		"task_identifier": taskIDStr,
 		"title":           ComputeTaskTitle(release),
 		"repo":            fmt.Sprintf("%s/%s", release.Repo.Owner, release.Repo.Name),
-		"clone_url":       fmt.Sprintf("git@github.com:%s/%s.git", release.Repo.Owner, release.Repo.Name),
+		"clone_url": fmt.Sprintf(
+			"git@github.com:%s/%s.git",
+			release.Repo.Owner,
+			release.Repo.Name,
+		),
 		"ref":             release.HeadSHA,
 		"current_version": release.CurrentVersion,
 	}
@@ -98,11 +102,15 @@ func buildTaskBody(release Release) string {
 	name := release.Repo.Name
 	return fmt.Sprintf(
 		"# Release: %s/%s\n\n**Current version:** %s\n**HEAD:** %s\n**Changelog:** https://github.com/%s/%s/blob/master/CHANGELOG.md\n**Repo:** [%s/%s](https://github.com/%s/%s)\n",
-		owner, name,
+		owner,
+		name,
 		release.CurrentVersion,
 		release.ShortSHA(),
-		owner, name,
-		owner, name,
-		owner, name,
+		owner,
+		name,
+		owner,
+		name,
+		owner,
+		name,
 	)
 }